Add files via upload
Course 5 - Regression Analysis: Simplify complex data relationships
You are a data professional at TikTok. The data team is working towards building a machine learning model that can be used to determine whether a video contains a claim or whether it offers an opinion. With a successful prediction model, TikTok can reduce the backlog of user reports and prioritize them more eﬀiciently.
The team is getting closer to completing the project, having completed an initial plan of action, initial Python coding work, EDA, and hypothesis testing.
The TikTok team has reviewed the results of the hypothesis testing. TikTok’s Operations Lead, Maika Abadi, is interested in how different variables are associated with whether a user is verified. Earlier, the data team observed that if a user is verified, they are much more likely to post opinions. Now, the data team has decided to explore how to predict verified status to help them understand how video characteristics relate to verified users. Therefore, you have been asked to conduct a logistic regression using verified status as the outcome variable. The results may be used to inform the final model related to predicting whether a video is a claim vs an opinion.

In this activity, you will build a logistic regression model in Python. As you have learned, logistic regression helps you estimate the probability of an outcome. For data science professionals, this is a useful skill because it allows you to consider more than one variable against the variable you’re measuring against. This opens the door for much more thorough and flexible analysis to be completed.

The purpose of this project is to demostrate knowledge of EDA and regression models.
The goal is to build a logistic regression model and evaluate the model. 

This activity has three parts:
Part 1: EDA & Checking Model Assumptions * What are some purposes of EDA before con- structing a logistic regression model?
Part 2: Model Building and Evaluation * What resources do you find yourself using as you complete this stage?
Part 3: Interpreting Model Results
• What key insights emerged from your model(s)?
• What business recommendations do you propose based on the models built?